Revert "moved the create users code to a reuseable lib to be able to be used elsewhere, updated tests to use roles instead of types"
This reverts commit 90471f041768cb9c33188b57b3eddbc58c25d8b0.

reverting changes to move to alib function. It was not needed for this change
