+++ test/data/create-users/existing-place/users.csv
@@ -1,2 +1,2 @@
-username,password,roles,name,phone,contact.c_prop,place
+username,password,type,name,phone,contact.c_prop,place
 todd,Secret_1,district-manager,Alice Example,+123456789,c_val_a,place_uuid_here
